pyproject.toml
[tool.poetry]
name = "mentor-match-web"
version = "3.0.1"
description = "A web service for matching mentors and mentees"
authors = ["jonathan.kerr <jonathan.kerr@digital.cabinet-office.gov.uk>"]
license = "Crown Open Source"

[tool.poetry.dependencies]
python = "^3.9"
amqp = "^5.1.1"
async_generator = "^1.10"
attrs = "^22.1.0"
"backports.entry-points-selectable" = "^1.2.0"
billiard = "^3.6.4"
black = "^22.10.0"
bump2version = "^1.0.1"
cachetools = "^5.2.0"
celery-stubs = "^0.1.2"
certifi = "^2022.9.24"
cffi = "^1.15.1"
cfgv = "^3.3.1"
charset-normalizer = "^3.0.0"
click = "^8.1.3"
click-didyoumean = "^0.3.0"
click-plugins = "^1.1.1"
click-repl = "^0.2.0"
commonmark = "^0.9.1"
coverage = "^6.5.0"
cryptography = "^38.0.1"
distlib = "^0.3.6"
docopt = "^0.6.2"
execnet = "^1.9.0"
extras = "^1.0.0"
filelock = "^3.8.0"
fixtures = "^4.0.1"
Flask = "^2.2.2"
freezegun = "^1.2.2"
fuzzywuzzy = "^0.18.0"
glob2 = "^0.7"
google-api-core = "^2.10.2"
google-api-python-client = "^2.65.0"
google-api-python-client-stubs = "^1.12.0"
google-auth = "^2.13.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.3"
googleapis-common-protos = "^1.56.4"
gunicorn = "^20.1.0"
h11 = "^0.14.0"
httplib2 = "^0.20.4"
identify = "^2.5.6"
iniconfig = "^1.1.1"
itsdangerous = "^2.1.2"
Jinja2 = "^3.1.2"
kombu = "^5.2.4"
lettuce = "^0.2.23"
Mako = "^1.2.3"
Markdown = "^3.4.1"
MarkupSafe = "^2.1.1"
mock = "^4.0.3"
munkres = "^1.1.4"
mypy = "^0.982"
mypy-extensions = "^0.4.3"
nodeenv = "^1.7.0"
notifications-python-client = "^6.4.1"
oauthlib = "^3.2.2"
outcome = "^1.2.0"
packaging = "^21.3"
parse = "^1.19.0"
parse-type = "^0.6.0"
pathspec = "^0.10.1"
pbr = "^5.11.0"
platformdirs = "^2.5.2"
pluggy = "^1.0.0"
pre-commit = "^2.20.0"
prompt-toolkit = "^3.0.31"
protobuf = "^4.21.8"
py = "^1.11.0"
pyasn1 = "^0.4.8"
pyasn1-modules = "^0.2.8"
pycparser = "^2.21"
Pygments = "^2.13.0"
PyJWT = "^2.6.0"
pyOpenSSL = "^22.1.0"
pyparsing = "^3.0.9"
python-dateutil = "^2.8.2"
python-subunit = "^1.4.0"
pytz = "^2022.5"
PyYAML = "^6.0"
redis = "^4.3.4"
requests-oauthlib = "^1.3.1"
rich = "^12.6.0"
rsa = "^4.9"
selenium = "^4.5.0"
six = "^1.16.0"
sniffio = "^1.3.0"
sortedcontainers = "^2.4.0"
sure = "^2.0.0"
testtools = "^2.5.0"
toml = "^0.10.2"
tomli = "^2.0.1"
tox = "^3.26.0"
trio = "^0.22.0"
trio-websocket = "^0.9.2"
types-freezegun = "^1.1.10"
types-mock = "^4.0.15"
types-requests = "^2.28.11"
types-urllib3 = "^1.26.25"
typing-extensions = "^4.4.0"
uritemplate = "^4.1.1"
urllib3 = "^1.26.12"
vine = "^5.0.0"
virtualenv = "^20.16.5"
wcwidth = "^0.2.5"
Werkzeug = "^2.2.2"
wsproto = "^1.2.0"
celery = "^5.2.7"
mentor-match = "^7.0.1"
aws-cdk-lib = "^2.69.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.test.dependencies]
pytest = "^7.1.3"
pytest-bdd = "^6.0.1"
pytest-cov = "^4.0.0"
pytest-forked = "^1.4.0"
pytest-xdist = "^2.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
